# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  projectId: string;
  project: string;
  portfolioFamily: string;
  tagsSecondaryFamilies: string;
  chatGptProjectName?: string;
  category: string;
  lifecycleStage: string;
  levelPercent: number;
  status: string;
  latestCompletedGate: string;
  currentGate: string;
  blocker: string;
  exactNextAction: string;
  doNotRedo: string;
  technicalStackProductType: string;
  currentWorkPlatform: string;
  betaStatus: string;
  launchStatus: string;
  lastActivity: string;
  lastTrackerUpdate: string;
  evidenceCheckpoint: string;
  notes: string;
  googleDriveFolder: string;
  documentationStatus: string;
}

export interface StageRule {
  lifecycleStage: string;
  levelPercent: number;
  gateRule: string;
}

export interface ChangeLogEntry {
  timestamp: string;
  project: string;
  previousStage: string;
  newStage: string;
  previousStatus: string;
  newStatus: string;
  gateChange: string;
  evidence: string;
  updatedBy: string;
}

export type PortfolioSnapshotMode = typeof PortfolioSnapshotMode[keyof typeof PortfolioSnapshotMode];


export const PortfolioSnapshotMode = {
  LIVE_DATA: 'LIVE DATA',
  SAMPLE_MODE: 'SAMPLE MODE',
} as const;

export interface PortfolioSnapshot {
  mode: PortfolioSnapshotMode;
  syncedAt: string;
  projects: Project[];
  changeLog: ChangeLogEntry[];
  stageRules: StageRule[];
  /** @nullable */
  warning?: string | null;
}

export interface PortfolioAssistantInput {
  /**
     * @minLength 1
     * @maxLength 2000
     */
  question: string;
}

export interface PortfolioAssistantProjectResult {
  projectId: string;
  project: string;
  portfolioFamily: string;
  tagsSecondaryFamilies: string;
  lifecycleStage: string;
  status: string;
  currentGate: string;
  blocker: string;
  exactNextAction: string;
}

export interface PortfolioAssistantEvidence {
  label: string;
  value: string;
}

export type PortfolioAssistantAnswerBasis = typeof PortfolioAssistantAnswerBasis[keyof typeof PortfolioAssistantAnswerBasis];


export const PortfolioAssistantAnswerBasis = {
  EXPLICIT_SHEET_DATA: 'EXPLICIT SHEET DATA',
  DERIVED_FROM_SHEET_DATA: 'DERIVED FROM SHEET DATA',
} as const;

export interface PortfolioAssistantAnswer {
  answer: string;
  basis: PortfolioAssistantAnswerBasis;
  usedAiInterpretation: boolean;
  projects: PortfolioAssistantProjectResult[];
  evidence: PortfolioAssistantEvidence[];
}

export interface ErrorResponse {
  error: string;
}