# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: portfolio
    description: Read-only founder portfolio operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /portfolio:
    get:
      operationId: getPortfolio
      tags: [portfolio]
      summary: Load the founder portfolio
      description: Returns the latest portfolio snapshot from Google Sheets or clearly labeled sample data when live credentials have never worked.
      responses:
        "200":
          description: Portfolio snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioSnapshot"
  /portfolio/projects/{projectId}:
    get:
      operationId: getPortfolioProject
      tags: [portfolio]
      summary: Load one project from the founder portfolio
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Project was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /portfolio/assistant:
    post:
      operationId: askPortfolioAssistant
      tags: [portfolio]
      summary: Ask the read-only portfolio assistant
      description: Answers from the latest Google Sheet-backed portfolio snapshot without changing project data.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioAssistantInput"
      responses:
        "200":
          description: Portfolio assistant answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioAssistantAnswer"
        "400":
          description: Invalid assistant question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Project:
      type: object
      required:
        - projectId
        - project
        - portfolioFamily
        - tagsSecondaryFamilies
        - category
        - lifecycleStage
        - levelPercent
        - status
        - latestCompletedGate
        - currentGate
        - blocker
        - exactNextAction
        - doNotRedo
        - technicalStackProductType
        - currentWorkPlatform
        - betaStatus
        - launchStatus
        - lastActivity
        - lastTrackerUpdate
        - evidenceCheckpoint
        - notes
        - googleDriveFolder
        - documentationStatus
      properties:
        projectId: { type: string }
        project: { type: string }
        portfolioFamily: { type: string }
        tagsSecondaryFamilies: { type: string }
        chatGptProjectName: { type: string }
        category: { type: string }
        lifecycleStage: { type: string }
        levelPercent: { type: number }
        status: { type: string }
        latestCompletedGate: { type: string }
        currentGate: { type: string }
        blocker: { type: string }
        exactNextAction: { type: string }
        doNotRedo: { type: string }
        technicalStackProductType: { type: string }
        currentWorkPlatform: { type: string }
        betaStatus: { type: string }
        launchStatus: { type: string }
        lastActivity: { type: string }
        lastTrackerUpdate: { type: string }
        evidenceCheckpoint: { type: string }
        notes: { type: string }
        googleDriveFolder: { type: string }
        documentationStatus: { type: string }
    StageRule:
      type: object
      required:
        - lifecycleStage
        - levelPercent
        - gateRule
      properties:
        lifecycleStage: { type: string }
        levelPercent: { type: number }
        gateRule: { type: string }
    ChangeLogEntry:
      type: object
      required:
        - timestamp
        - project
        - previousStage
        - newStage
        - previousStatus
        - newStatus
        - gateChange
        - evidence
        - updatedBy
      properties:
        timestamp: { type: string }
        project: { type: string }
        previousStage: { type: string }
        newStage: { type: string }
        previousStatus: { type: string }
        newStatus: { type: string }
        gateChange: { type: string }
        evidence: { type: string }
        updatedBy: { type: string }
    PortfolioSnapshot:
      type: object
      required:
        - mode
        - syncedAt
        - projects
        - changeLog
        - stageRules
      properties:
        mode:
          type: string
          enum: [LIVE DATA, SAMPLE MODE]
        syncedAt: { type: string }
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        changeLog:
          type: array
          items:
            $ref: "#/components/schemas/ChangeLogEntry"
        stageRules:
          type: array
          items:
            $ref: "#/components/schemas/StageRule"
        warning: { type: ["string", "null"] }
    PortfolioAssistantInput:
      type: object
      required:
        - question
      properties:
        question:
          type: string
          minLength: 1
          maxLength: 2000
    PortfolioAssistantProjectResult:
      type: object
      required:
        - projectId
        - project
        - portfolioFamily
        - tagsSecondaryFamilies
        - lifecycleStage
        - status
        - currentGate
        - blocker
        - exactNextAction
      properties:
        projectId: { type: string }
        project: { type: string }
        portfolioFamily: { type: string }
        tagsSecondaryFamilies: { type: string }
        lifecycleStage: { type: string }
        status: { type: string }
        currentGate: { type: string }
        blocker: { type: string }
        exactNextAction: { type: string }
    PortfolioAssistantEvidence:
      type: object
      required:
        - label
        - value
      properties:
        label: { type: string }
        value: { type: string }
    PortfolioAssistantAnswer:
      type: object
      required:
        - answer
        - basis
        - usedAiInterpretation
        - projects
        - evidence
      properties:
        answer: { type: string }
        basis:
          type: string
          enum: [EXPLICIT SHEET DATA, DERIVED FROM SHEET DATA]
        usedAiInterpretation: { type: boolean }
        projects:
          type: array
          items:
            $ref: "#/components/schemas/PortfolioAssistantProjectResult"
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/PortfolioAssistantEvidence"
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }