# lib/api-zod/src/generated/types/portfolioAssistantAnswer.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PortfolioAssistantAnswerBasis } from './portfolioAssistantAnswerBasis';
import type { PortfolioAssistantEvidence } from './portfolioAssistantEvidence';
import type { PortfolioAssistantProjectResult } from './portfolioAssistantProjectResult';

export interface PortfolioAssistantAnswer {
  answer: string;
  basis: PortfolioAssistantAnswerBasis;
  usedAiInterpretation: boolean;
  projects: PortfolioAssistantProjectResult[];
  evidence: PortfolioAssistantEvidence[];
}