# lib/api-zod/src/generated/types/portfolioAssistantInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PortfolioAssistantInput {
  /**
     * @minLength 1
     * @maxLength 2000
     */
  question: string;
}