# lib/api-zod/src/generated/types/portfolioAssistantProjectResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PortfolioAssistantProjectResult {
  projectId: string;
  project: string;
  portfolioFamily: string;
  tagsSecondaryFamilies: string;
  lifecycleStage: string;
  status: string;
  currentGate: string;
  blocker: string;
  exactNextAction: string;
}