# lib/api-zod/src/generated/types/portfolioSnapshot.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChangeLogEntry } from './changeLogEntry';
import type { PortfolioSnapshotMode } from './portfolioSnapshotMode';
import type { Project } from './project';
import type { StageRule } from './stageRule';

export interface PortfolioSnapshot {
  mode: PortfolioSnapshotMode;
  syncedAt: string;
  projects: Project[];
  changeLog: ChangeLogEntry[];
  stageRules: StageRule[];
  /** @nullable */
  warning?: string | null;
}