# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Project {
  projectId: string;
  project: string;
  portfolioFamily: string;
  tagsSecondaryFamilies: string;
  chatGptProjectName?: string;
  category: string;
  lifecycleStage: string;
  levelPercent: number;
  status: string;
  latestCompletedGate: string;
  currentGate: string;
  blocker: string;
  exactNextAction: string;
  doNotRedo: string;
  technicalStackProductType: string;
  currentWorkPlatform: string;
  betaStatus: string;
  launchStatus: string;
  lastActivity: string;
  lastTrackerUpdate: string;
  evidenceCheckpoint: string;
  notes: string;
  googleDriveFolder: string;
  documentationStatus: string;
}