# lib/api-zod/src/generated/types/stageRule.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StageRule {
  lifecycleStage: string;
  levelPercent: number;
  gateRule: string;
}